SolidState

Documentation for SolidState.

SolidState.ChartInfo

SolidState.ChartInfo

SolidState.ChartType

SolidState.DataChart

SolidState.DataChart

SolidState.DataIntegral

SolidState.DataIntegral

SolidState.DataIntegral

SolidState.DataIntegral

SolidState.DataIntegral

SolidState.DataIntegral

SolidState.DataMap

SolidState.DataMap

SolidState.DataMap

SolidState.DataSection

SolidState.DataSection

SolidState.DataSection

SolidState.DataSection

SolidState.RangeType

SolidState.VOL

SolidState.LatticePoints

SolidState.R2D

SolidState.base_ranges

SolidState.class_weights

SolidState.cointegrate

SolidState.combine_bases

SolidState.data_export

SolidState.data_import

SolidState.edge_test

SolidState.integral_data

SolidState.integrate

SolidState.integrate

SolidState.integrate

SolidState.integrate!

SolidState.prior_base_combine

SolidState.square_integration_weights

SolidState.strip_divide

SolidState.weight_classify

SolidState.RangeType — Type

const RangeType = Union{ Tuple{Symbol,NTuple{N,T}} where {N, T <: AbstractRange}, Tuple{Symbol,AV} where AV <: AbstractVector, Vector{Tuple{Symbol,N,N,I}}, Tuple{Symbol,Symbol,I}, Vector{NTuple{NT,I}}} where {N <: Number, I <: Integer, NT }

SolidState.ChartInfo — Type

struct ChartInfo{T <: ChartType, RT1 <: RangeType, RT2 <: RangeType, RT3 <: RangeType, CAT <: Union{Tuple{Int,Int}}} type::Type{T} indices::RT1 priors::RT2 base::RT3 mn::CAT end

SolidState.ChartInfo — Method

function ChartInfo() ChartInfo(ChartType,[(1,)],[(1,)],[(1,)],(0,0)) end

SolidState.ChartType — Type

abstract type ChartType end

SolidState.DataChart — Type

DataChart{IndexT <: AbstractArray, PriorT <: AbstractArray, BaseT <: AbstractArray, DataT <: AbstractArray}

A generic structure to hold Tensor Data

SolidState.DataChart — Method

DataChart(Indices::AbstractArray, Priors::AbstractArray, Base::AbstractArray, DType::DataType; style=:normal)

SolidState.DataIntegral — Type

DataIntegral(ci::ChartInfo, cachedir::String; a::AbstractVector=[0.0,0.0], b::AbstractVector =[1.0,1.0])::DataIntegral

SolidState.DataIntegral — Type

DataIntegral{DM <: DataMap, DA <: AbstractArray, AV <: AbstractVector, IA <: AbstractArray, EA <: AbstractArray, CIT <: ChartInfo}

SolidState.DataIntegral — Method

DataIntegral(asd, mn, datatype::Type{T} where T <: ChartType,indices,priors, base, cachedir=ENV["cachedir"], a::AV=[0.0,0.0], b::AV =[1.0,1.0])::DataIntegral  where {AV <: AbstractVector}

Construction with cached objects & description

SolidState.DataIntegral — Method

DataIntegral(datatype::Type{T} where T <: ChartType,asd::Dict{String,Any},hd::HamiltonianDensity,indices,priors,base; a::AV=[0.0,0.0], b::AV =[1.0,1.0])::DataIntegral  where {AV <: AbstractVector, DM <: DataMap}

SolidState.DataIntegral — Method

DataIntegral(dm::DM, a::AV=[0.0,0.0], b::AV=[1.0,1.0];cachedir::String="none", ranges::ChartInfo=ChartInfo())::DataIntegral where {AV <: AbstractVector, DM <: DataMap, PT <: Union{AbstractWorkerPool,Symbol}}

SolidState.DataIntegral — Method

DataIntegral(; asd, dtype::Type{T} where T <: ChartType,indices,priors, base, cachedir, mn, a::AV=[0.0,0.0], b::AV =[1.0,1.0], kargs...)::DataIntegral  where {AV <: AbstractVector}

Construction with cached objects & description

SolidState.DataMap — Method

DataMap(pmodel,pchart; cachedir=ENV["cachedir"], Λ0=:auto, style=:normal, kargs...)

SolidState.DataMap — Method

DataMap(model::Symbol, dtype::(Type{T} where T <: ChartType), indices0, prange, brange; cachedir, mn, Λ0=:auto, style=:normal)

Uses cached model

SolidState.DataMap — Method

DataMap(; asd, dtype::(Type{T} where T <: ChartType), indices0, prange, brange, cachedir, mn, Λ0=:auto, style=:normal)

Uses cached model

SolidState.DataSection — Type

DataSection{DM <: DataMap, TC <: DataChart}

A structure to hold sampling information of DataMaps over the sampling base space

SolidState.DataSection — Method

DataSection(dm::DataMap, dom::AbstractArray)

DataSection constructor

SolidState.DataSection — Method

DataSection(dm::DataMap, dom::Vector{Tuple{Symbol,N,N,Int64}} where N <: Number)

A Data Section constructor with a range input

SolidState.DataSection — Method

DataSection(;asd,mn,dtype,N,indices,priors,base,cachedir,offset, kargs...)

Dictionary based constructor

SolidState.VOL — Type

VOL

SolidState.LatticePoints — Method

LatticePoints(xtal::Tuple{Array{Float64,2},Vector{Vector{T}}} where T <: Real, bnds::Array{Int64,2})::Vector{Vector{Vector{Float64}}}

SolidState.R2D — Method

function R2D(α::Real,θ::Real)::Matrix{Float64}

SolidState.base_ranges — Method

base_ranges(ds::DataSection, i_b::Int)

SolidState.class_weights — Method

class_weights(w::Tuple{Symbol,Symbol})

Returns symmetrized square triangular integral mesh weights. The weight corresponds to the average over global triangularization schemes and give corners 1.5, edges 3, and bulk 6.

SolidState.cointegrate — Method

cointegrate(ci::ChartInfo,evals,cachedir,a,b)

SolidState.combine_bases — Method

combine_bases(oldbase,dombase)

SolidState.data_export — Method

data_export(datadir,strct)

SolidState.data_import — Method

data_import(datadir)

SolidState.edge_test — Method

edge_test(x::NTuple{N, Float64} where N ,mins::AbstractArray,maxes::AbstractArray)

SolidState.integral_data — Method

integral_data(ds::DataSection)

Returns the integral in an AbstractArray type that corresponds to that of the data in DataMap of ds.

SolidState.integrate! — Method

integrate!(di::DataIntegral, ds::DataSection)

In place push of data section integral to the passed DataIntegral

SolidState.integrate — Method

integrate(dm::DM where DM <: DataMap, a::AbstractVector, b::AbstractVector, pool::AbstractWorkerPool; rtol=1e-12, atol=1e-12, evals::AbstractRange, ranges::ChartInfo=ChartInfo(), cachedir::String)

SolidState.integrate — Method

integrate(dm::DM where DM <: DataMap, a::AbstractVector, b::AbstractVector, pool::Symbol = :none; rtol=1e-12, atol=1e-12, evals::Int=typemax(Int), kwargs...)::Float64

Core Call to HCubature

SolidState.integrate — Method

integrate(ds::DataSection)

Returns DataIntegral with identical DataMap and the output of integrate(ds) pushed into the first data position

SolidState.prior_base_combine — Method

prior_base_combine(oldpriors,oldbase)

SolidState.square_integration_weights — Method

square_integration_weights(base)

Returns the weights associated to the base of a data section

SolidState.strip_divide — Method

function strip_divide(a,b,nw)

For parallel Integration, to divide hyper k-cube into (k-1) slabs. cube is defined between to points a & b. nw specifies the number of workers, i.e. strips

SolidState.weight_classify — Method

weight_classify(x::NTuple{N, Float64} where N ,mins::AbstractArray,maxes::AbstractArray,i::Int)

Return a tuple classifying the location of each coordinate in the grid